% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EbmPanelData-package.R
\docType{package}
\name{EbmPanelData-package}
\alias{EbmPanelData}
\alias{EbmPanelData-package}
\title{EbmPanelData: Panel Data Regression with Fixed Effects}
\description{
Functions connected to fixed effect models for panel data.
}
\author{
\strong{Maintainer}: Christof Schötz \email{cschoetz@pik-potsdam.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
